requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
requests==2.31.0

# AI/ML Dependencies
# Optional AI/ML dependencies removed for lightweight local dev. Add them back if you want full evaluator.
# sentence-transformers==2.2.2
# torch==2.2.2
# nltk==3.8.1
# numpy==1.24.3
# scikit-learn==1.3.2

# File handling
aiofiles==23.2.0